Courseware/Labs/10_Delegates_and_Lambdas/01 Pizza/Begin/PizzaProject/Checkout/Checkout.cs
using System;
using System.Collections.Generic;
using System.Text;

namespace PizzaProject
{
    public class Checkout
    {
        readonly BestDiscount bestDiscount;
        public Checkout(BestDiscount bestDiscount)
        {
            this.bestDiscount = bestDiscount;
        }

        public PriceData GetBestPrice(Order order)
        {
            DiscountPolicyData discountPolicyData = bestDiscount.GetBestDiscount(order);
            PriceData priceData = new PriceData(discountPolicyData.DiscountPolicyName, order.NonDiscountedPrice, discountPolicyData.Discount);
            return priceData;
        }
    }
}
